Deduplicate ClosedStream types
